Fixed broken links from merge
Addressed merge conflicts and missing file references in response to the reorganization of scripts between the main and data-analytics branch
