+++ Assets/Input/PlayerController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5cfccc29b98a78b4d9e59c3a03a45afa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
